Add deprecation to rename a class being generic now
